Fix usdhc.disk_usdhc_access_write busy fail
For unknown reasons, calling disk_usdhc_access_write continuously
will check that the CDIHB field of the register PRES_STATE is 1
after the previous transfer is completed and before the next transfer.

This phenomenon is more likely to occur. May be a SOC issue, and more
experiments are needed to confirm.

This is a temporary solution, wait for CDIHB to become 0 before
transmitting data.

Signed-off-by: Frank Li <lgl88911@163.com>